--- app/(components)/index.ts
@@ -1,2 +1,0 @@
-export * from './layout';
-export * from './scripts';
