+++ app/_components/CookieConsent/scripts/index.ts
