+++ app/_components/Header/Header.tsx
@@ -7,7 +7,7 @@ import dynamic from 'next/dynamic';
 import Link from 'next/link';
 import { useIntl } from 'react-intl';
 
-import { useTheme } from '@/app/(contexts)/theme';
+import { useTheme } from '@/app/_contexts/theme';
 import { GithubIcon } from '@/components/icons';
 import {
   Button,
