+++ src/constants/cookies.ts
@@ -1,5 +1,6 @@
 export const PREFIX = 'juniors_bootcamp';
 export const COOKIES = {
   THEME: `${PREFIX}_theme`,
-  COOKIE_CONSENT: `${PREFIX}_cookie_consent`
+  COOKIE_CONSENT: `${PREFIX}_cookie_consent`,
+  USER_SETTINGS: `${PREFIX}_user_settings`
 } as const;
